Django_hw/pyproject.toml
[tool.poetry]
name = "django_hw_13"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_hw_13"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
django-taggit = "^3.1.0"
pymongo = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
